Span issue 9 (#14)
* Changed fragment from Int? to Optional<Int>

As per span issue #9 . Justified in that fragment size == auto is not
missing or invalid data (usual meaning of null).

* Fixed tests based on fit information de/serialization

Had to bump the fit information file version to 2

* Fixed tests based on fragment size logging

* Replaced Optional<Int> with a dedicated sealed class Fragment

This should be easier to read and to maintain

* Fixed an error in Gson de/serialization

Also removed the [SPAN} link in kotlin doc which was always causing
a warning

* Massive refactoring of SpanCLA option parsing and interpretation

Part of work on span issue #9. If the --model option is supplied
and the corresponding path exists, it eliminates the need for
most of the other arguments. If those arguments are supplied anyway,
they are checked against those inferred from the model.

* Correct keep-dup option processing

Duplicates can now be explicitly forbidden with --keep-dup false

* Fixed some tests failing after SpanCLA refactoring

* More correct log order in SpanCLA

* SpanModelFitResults.genomeQuery no longer fails

Now Genome can function without a chrom.sizes file, so we just load
the chromosome size map from the model

* Fixed command-line parameter parsing in SpanCLA

Also extracted common functionality and added default values
to the help message

* Fixed SpanCLA "compare" option parsing

* Edited a suspicious requireNotNull line in Labels.kt

* Expanded model loading test in SpanCLALongTest

* Updated span CHANGES.txt

* Fixed a test in SpanCLALongTest

We've reduced the maximum number of Span iterations